Update SSL handling code for Python 3.13
The old code relied on a top-level "wrap_socket" static method
in the ssl module. This has apparently moved under the ssl.SSLContext
class. Update code to use an SSLContext with the appropriate settings
to function.

Note that we're not stressing too hard with properly verifying a
valid SSL cert. The included one is expired, and was just included
to make passive network monitoring in a CTF style environment not
useful for getting the key if someone else solves the challenge.